Extracting enemy flip, instantiation from HelmetFish to Enemy class. Extracting "Grounded" logic to LivingThing class. Alien.

+++ Assets/Scripts/HealthBarScript.cs
@@ -7,12 +7,12 @@ public class HealthBarScript : MonoBehaviour
 {
     public Slider slider;
 
-    public void setMaxHealth(int health){
+    public void setMaxHealth(float health){
         slider.maxValue = health;
         slider.value = health;
     }
 
-    public void SetHealth(int health){
+    public void SetHealth(float health){
         slider.value = health;
     }
 }
